requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.2
bcrypt==4.2.1
blinker==1.9.0
blis==1.2.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.37.3
botocore==1.37.3
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
edge-tts==7.0.0
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.2.0
fabric==3.2.2
ffmpeg==1.4
ffprobe==0.5
filelock==3.17.0
fish-audio-sdk==2025.2.11
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
g2pkk==0.1.2
Gentle==0.1
gotrue==2.11.4
groq==0.18.0
grpcio==1.70.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.14.0
h2==4.2.0
hangul-romanize==0.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.1
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
invoke==2.2.0
ipython==8.32.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
ormsgpack==1.8.0
packaging==24.2
pandas==1.5.3
paramiko==3.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
postgrest==0.19.3
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.2.1
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3transfer==0.11.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.11.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
srt==3.5.3
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
SudachiDict-core==20250129
SudachiPy==0.6.10
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.49.0
TTS==0.22.0
typeguard==4.4.2
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.3
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
